Drop support for older YunoHost versions.
I'm not sure how long it will take for people to migrate on YunoHost 11.
Untill then, the package will not be available so bad news for them...
Even if package should work on it, I cannot test against it as I don't
use version 4.X anymore. Is there a CI for it?